+++ CHANGELOG.md
@@ -676,5 +676,7 @@ This major version introduces the concept of `Engines` which handle the chunking
 
 ## v1.2.1
 
+- Outputs:
+  - Output `FarmResultsEval` now makes a difference between the states contraction rules `weights` and `mean`, make sure you adopt your code accordingly (usually you would want to use `mean`).
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.2.1](https://github.com/FraunhoferIWES/foxes/commits/v1.2.1)